chore: Move the ors config helper
There will be more config helper. It makes sense to isolate them
