+++ ors-test-scenarios/src/test/java/integrationtests/ConfigProfileTest.java
@@ -21,7 +21,7 @@ import java.util.Map;
 
 import static org.testcontainers.utility.MountableFile.forHostPath;
 import static utils.ContainerInitializer.initContainer;
-import static utils.OrsConfigHelper.configWithCustomProfilesActivated;
+import static utils.configs.OrsConfigHelper.configWithCustomProfilesActivated;
 
 @ExtendWith(TestcontainersExtension.class)
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
